pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "PictSure"
version = "0.1.0"
description = "A package for generalized image classification using In-Context-Learning with PyTorch."
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Cornelius Wolff", email = "cornelius.wolff@cwi.nl" },
    { name = "Lukas Schiesser", email = "lukas.schiesser@dfki.de" }
]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "numpy>=1.26.4",
    "Pillow",
    "click>=8.1.7",
    "tqdm>=4.66.4",
    "requests>=2.32.3",
    "huggingface-hub>=0.33.1"
]

[project.scripts]
pictsure = "PictSure.cli:cli"

[tool.setuptools.packages.find]